Refactor map controls panel widgets
